Organized files in unit tests project.
