+++ RandomizerUnitTests/Editor/CompareGameObjectProperties.cs
