+++ RandomizerUnitTests/Editor/TestGameObjectEquality.cs
