+++ RandomizerUnitTests/Editor/TestUWLinkedList.cs
@@ -18,7 +18,7 @@ public class TestUWLinkedList
     private List<GameObject> _gameObjects;
     private UWLinkedList LList1;
     private UWLinkedList LList2;
-    
+
     [SetUp]
     public void Setup()
     {
@@ -36,8 +36,6 @@ public class TestUWLinkedList
 
         LList1 = new UWLinkedList();
         LList2 = new UWLinkedList();
-
-
     }
 
     [Test]
@@ -51,7 +49,7 @@ public class TestUWLinkedList
         LList1.startingIdx = 3; // This should clear the list
         Assert.True(LList1.Count == 0);
         LList1.PopulateObjectList(_gameObjects.ToArray());
-        
+
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {3, 4, 5}));
         Assert.True(LList1.Count == 3);
 
@@ -59,14 +57,13 @@ public class TestUWLinkedList
         Assert.True(LList1.Count == 0);
         LList1.PopulateObjectList(_gameObjects.ToArray());
         Assert.True(LList1.Count == 0);
-        
+
         // Trying to trigger the ContraintException
         _gameObjects[1].next = 2;
         _gameObjects[2].next = 1;
         LList1.Clear();
         LList1.startingIdx = 1;
         Assert.Throws<ConstraintException>(() => LList1.PopulateObjectList(_gameObjects.ToArray()));
-
     }
 
     [Test]
@@ -76,7 +73,7 @@ public class TestUWLinkedList
         LList2.startingIdx = 0;
         LList1.PopulateObjectList(_gameObjects);
         LList2.PopulateObjectList(_gameObjects);
-        
+
         // Assert.Throws<InvalidOperationException>(() => throw new InvalidOperationException());
         Assert.True(LList1.Pop().Equals(_gameObjects[6]));
         Assert.True(LList1.Pop().Equals(_gameObjects[2]));
@@ -84,7 +81,7 @@ public class TestUWLinkedList
         Assert.Throws<InvalidOperationException>(() => LList1.Pop());
         Assert.Throws<InvalidOperationException>(() => LList2.Pop());
     }
-    
+
     [Test]
     public void TestPopulateObjectList_withList()
     {
@@ -99,24 +96,24 @@ public class TestUWLinkedList
     {
         LList1.startingIdx = 1;
         LList2.startingIdx = 3;
-        
+
         LList1.PopulateObjectList(_gameObjects.ToArray());
         LList2.PopulateObjectList(_gameObjects.ToArray());
-        
+
         // LList1 has items in idx 1*,2,6*
         // LList2 has items in idx 3,4,5*
         // * means it should be moved
         var objs1 = LList1.PopObjectsThatShouldBeMoved();
         var objs2 = LList2.PopObjectsThatShouldBeMoved();
-        
+
         // Now LList1 should be 2, LList2 should be 3,4
         // objs1 should be 1,6
         // objs2 should be 5
-        Assert.True(LList1.Count + objs1.Count == 3);  // Should be complementary
+        Assert.True(LList1.Count + objs1.Count == 3); // Should be complementary
         Assert.True(objs1[0].Equals(_gameObjects[1]));
         Assert.True(LList1[0].Equals(_gameObjects[2]));
         Assert.True(objs1[1].Equals(_gameObjects[6]));
-        
+
         Assert.True(LList2.Count + objs2.Count == 3);
         Assert.True(LList2[0].Equals(_gameObjects[3]));
         Assert.True(LList2[1].Equals(_gameObjects[4]));
@@ -135,53 +132,53 @@ public class TestUWLinkedList
     {
         LList1.startingIdx = 1;
         LList2.startingIdx = 3;
-        
+
         // LList1 has items in idx 1*,2,6*
         // LList2 has items in idx 3,4,5*
         // * means it should be moved
         LList1.PopulateObjectList(_gameObjects.ToArray());
         LList2.PopulateObjectList(_gameObjects.ToArray());
-        
+
         // Now LList1 should be 2, LList2 should be 3,4
         // objs1 should be 1,6
         // objs2 should be 5
         var objs1 = LList1.PopObjectsThatShouldBeMoved();
         var objs2 = LList2.PopObjectsThatShouldBeMoved();
-        
+
         // LList1 should have 2,5
         // LList2 should have 3,4,1,6
         LList1.AppendItems(objs2.ToArray()); // Changing to ToArray just to cover the test
         LList2.AppendItems(objs1.ToArray());
-        
+
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {2, 5}));
         Assert.True(CheckObjectsAtPositions(LList2.ToList(), new List<short>() {3, 4, 1, 6}));
         Assert.True(LList1.CheckIntegrity());
         Assert.True(LList2.CheckIntegrity());
     }
-    
+
     [Test]
     public void TestAppendItems_OnEmpty()
     {
         LList1.startingIdx = 1;
         LList2.startingIdx = 0;
-        
+
         // LList1 has items in idx 1*,2,6*
         // LList2 has no items
         // * means it should be moved
         LList1.PopulateObjectList(_gameObjects.ToArray());
         LList2.PopulateObjectList(_gameObjects.ToArray());
-        
+
         // Now LList1 should be 2, LList2 should be empty
         // objs1 should be 1,6
         // objs2 should be empty
         var objs1 = LList1.PopObjectsThatShouldBeMoved();
         var objs2 = LList2.PopObjectsThatShouldBeMoved();
-        
+
         // LList1 should have 2
         // LList2 should have 1,6
         LList1.AppendItems(objs2);
         LList2.AppendItems(objs1);
-        
+
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {2}));
         Assert.True(CheckObjectsAtPositions(LList2.ToList(), new List<short>() {1, 6}));
         Assert.True(LList1.CheckIntegrity());
@@ -193,24 +190,24 @@ public class TestUWLinkedList
     {
         LList1.startingIdx = 1;
         LList2.startingIdx = 3;
-        
+
         // LList1 has items in idx 1*,2,6*
         // LList2 has items in idx 3,4,5*
         // * means it should be moved
         LList1.PopulateObjectList(_gameObjects.ToArray());
         LList2.PopulateObjectList(_gameObjects.ToArray());
-        
+
         // Now LList1 should be 2, LList2 should be 3,4
         // objs1 should be 1,6
         // objs2 should be 5
         var objs1 = LList1.PopObjectsThatShouldBeMoved();
         var objs2 = LList2.PopObjectsThatShouldBeMoved();
-        
+
         // LList1 should have 5,2
         // LList2 should have 1,6,3,4
         LList1.PrependItems(objs2);
         LList2.PrependItems(objs1);
-        
+
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {5, 2}));
         Assert.True(CheckObjectsAtPositions(LList2.ToList(), new List<short>() {1, 6, 3, 4}));
         Assert.False(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {2, 5}));
@@ -218,30 +215,30 @@ public class TestUWLinkedList
         Assert.True(LList1.CheckIntegrity());
         Assert.True(LList2.CheckIntegrity());
     }
-    
+
     [Test]
     public void TestPrependItems_OnEmptyLList()
     {
         LList1.startingIdx = 1;
         LList2.startingIdx = 0;
-        
+
         // LList1 has items in idx 1*,2,6*
         // LList2 has no items
         // * means it should be moved
         LList1.PopulateObjectList(_gameObjects.ToArray());
         LList2.PopulateObjectList(_gameObjects.ToArray());
-        
+
         // Now LList1 should be 2, LList2 should be empty
         // objs1 should be 1,6
         // objs2 should be empty
         var objs1 = LList1.PopObjectsThatShouldBeMoved();
         var objs2 = LList2.PopObjectsThatShouldBeMoved();
-        
+
         // LList1 should have 2
         // LList2 should have 1,6
         LList1.PrependItems(objs2);
         LList2.PrependItems(objs1);
-        
+
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {2}));
         Assert.True(CheckObjectsAtPositions(LList2.ToList(), new List<short>() {1, 6}));
         Assert.True(LList1.CheckIntegrity());
@@ -276,23 +273,22 @@ public class TestUWLinkedList
         Assert.Throws<IndexOutOfRangeException>(() => LList1.RemoveAt(0));
         Assert.True(LList1.startingIdx == 0);
         Assert.True(LList1.CheckIntegrity());
-
     }
-    
-    [Test]
-    public void TestRemoveAt_FrontToBack(){
 
+    [Test]
+    public void TestRemoveAt_FrontToBack()
+    {
         // LList1 has items in idx 1,2,6
         LList1.startingIdx = 1;
         LList1.PopulateObjectList(_gameObjects);
-        
+
         LList1.RemoveAt(0); // items 2,6
         Assert.True(LList1.Count == 2);
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {2, 6}));
         Assert.Throws<IndexOutOfRangeException>(() => LList1.RemoveAt(2));
         Assert.True(LList1.startingIdx == LList1[0].IdxAtObjectArray);
         Assert.True(LList1.CheckIntegrity());
-        
+
         LList1.RemoveAt(0); // items 6
         Assert.True(LList1.Count == 1);
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {6}));
@@ -300,19 +296,19 @@ public class TestUWLinkedList
         Assert.True(LList1.startingIdx == LList1[0].IdxAtObjectArray);
         Assert.True(LList1.CheckIntegrity());
     }
-    
-    [Test]
-    public void TestRemoveAt_Middle(){
 
+    [Test]
+    public void TestRemoveAt_Middle()
+    {
         // LList1 has items in idx 1,2,6
         LList1.startingIdx = 1;
         LList1.PopulateObjectList(_gameObjects);
-        
+
         LList1.RemoveAt(1); // items 1,6
         Assert.True(LList1.Count == 2);
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {1, 6}));
         Assert.True(LList1.CheckIntegrity());
-        
+
         LList1.RemoveAt(1); // items 1
         Assert.True(LList1.Count == 1);
         Assert.True(CheckObjectsAtPositions(LList1.ToList(), new List<short>() {1}));
@@ -359,19 +355,19 @@ public class TestUWLinkedList
     {
         LList1.startingIdx = 1;
         LList1.PopulateObjectList(_gameObjects);
-        
+
         LList1.Insert(0, _gameObjects[3]);
         Assert.True(LList1.Count == 4);
         Assert.True(LList1.CheckIntegrity());
         LList1.RemoveAt(0);
         Assert.True(LList1.CheckIntegrity());
-        
+
         LList1.Insert(3, _gameObjects[3]);
         Assert.True(LList1.Count == 4);
         Assert.True(LList1.CheckIntegrity());
         LList1.RemoveAt(3);
         Assert.True(LList1.CheckIntegrity());
-        
+
         LList1.Insert(2, _gameObjects[3]);
         Assert.True(LList1.Count == 4);
         Assert.True(LList1.CheckIntegrity());
@@ -385,10 +381,12 @@ public class TestUWLinkedList
     {
         LList1.startingIdx = 1;
         LList1.PopulateObjectList(_gameObjects);
-        foreach (var foo in LList1) { }
+        foreach (var foo in LList1)
+        {
+        }
         // todo: How do I get to IEnumerator IEnumerable.GetEnumerator?
     }
-    
+
 
     private bool CheckObjectsAtPositions(List<GameObject> list, List<short> correctIdxs)
     {
@@ -396,6 +394,7 @@ public class TestUWLinkedList
         {
             throw new InvalidOperationException("Can't compare two lists of unequal counts!");
         }
+
         for (int i = 0; i < list.Count; i++)
         {
             if (!_gameObjects[correctIdxs[i]].Equals(list[i]))
@@ -414,33 +413,32 @@ public class TestUWLinkedList
         LList1.startingIdx = 1;
         LList1.PopulateObjectList(_gameObjects);
         Assert.True(LList1.CheckIntegrity());
-        
+
         // Changing the index of the first object should invalidate the list because it wouldn't be pointing correctly
         LList1[0].IdxAtObjectArray = 100; // Should be 1
         Assert.False(LList1.CheckIntegrity());
         LList1[0].IdxAtObjectArray = 1; // Returning to normal
         Assert.True(LList1.CheckIntegrity());
-        
+
         // Changing the index of an object in the middle, or its next attribute, should invalidate the LList
         LList1[1].IdxAtObjectArray = 100; // Should be 2
         Assert.False(LList1.CheckIntegrity());
         LList1[1].IdxAtObjectArray = 2; // Returning
         Assert.True(LList1.CheckIntegrity());
-        
+
         LList1[1].next = 100; // Should be 6
         Assert.False(LList1.CheckIntegrity());
         LList1[1].next = 6; // Returning
         Assert.True(LList1.CheckIntegrity());
-        
+
         // 
         _gameObjects[^1].next = 100;
-        Assert.False(LList1.CheckIntegrity()); 
-        
+        Assert.False(LList1.CheckIntegrity());
+
         // Having a starting index but no objects is invalid
         LList1.Clear();
         LList1.startingIdx = 1;
         Assert.False(LList1.CheckIntegrity());
-        
     }
 
     [Test]
@@ -479,7 +477,6 @@ public class TestUWLinkedList
         var LList4 = new UWLinkedList(LList1.ToArray(), (short) LList1.startingIdx);
         Assert.True(LList3.CheckIntegrity());
         Assert.True(LList4.CheckIntegrity());
-        
     }
 
     [Test]
@@ -494,10 +491,10 @@ public class TestUWLinkedList
         Assert.True(LList1.CheckIntegrity());
         LList1[1] = _gameObjects[5];
         Assert.True(LList1.CheckIntegrity());
-        
+
         Assert.Throws<IndexOutOfRangeException>(() => LList1[-1] = _gameObjects[2]);
         Assert.Throws<IndexOutOfRangeException>(() => LList1[100] = _gameObjects[2]);
-        
+
         LList1.Clear();
         LList1.Add(_gameObjects[1]);
         LList1[0] = _gameObjects[2];
@@ -509,7 +506,7 @@ public class TestUWLinkedList
     {
         LList1.startingIdx = 1;
         LList1.PopulateObjectList(_gameObjects);
-        
+
         Assert.True(LList1[0].Equals(_gameObjects[1]));
         Assert.True(LList1[1].Equals(_gameObjects[2]));
         Assert.True(LList1[2].Equals(_gameObjects[6]));
